refactor(inspectors): separated inspectors in own files
