+++ src/helpers/data.js
@@ -1,0 +1,41 @@
+function _toByteArray (hexString) {
+  const outArray = [];
+  const byteSize = 2;
+  for (let i = 0; i < hexString.length; i += byteSize) {
+    outArray.push(parseInt(hexString.substring(i, i + byteSize), 16));
+  }
+  return outArray;
+}
+
+function _toUTF8Data (string) {
+  const utf8 = [];
+  for (let i = 0; i < string.length; i++) {
+    let charcode = string.charCodeAt(i);
+    if (charcode < 0x80) {
+      utf8.push(charcode);
+    } else if (charcode < 0x800) {
+      utf8.push(0xc0 | (charcode >> 6), 0x80 | (charcode & 0x3f));
+    } else if (charcode < 0xd800 || charcode >= 0xe000) {
+      utf8.push(
+        0xe0 | (charcode >> 12),
+        0x80 | ((charcode >> 6) & 0x3f),
+        0x80 | (charcode & 0x3f)
+      );
+    } else {
+      // surrogate pair
+      i++;
+      // UTF-16 encodes 0x10000-0x10FFFF by
+      // subtracting 0x10000 and splitting the
+      // 20 bits of 0x0-0xFFFFF into two halves
+      charcode =
+        0x10000 + (((charcode & 0x3ff) << 10) | (string.charCodeAt(i) & 0x3ff));
+      utf8.push(
+        0xf0 | (charcode >> 18),
+        0x80 | ((charcode >> 12) & 0x3f),
+        0x80 | ((charcode >> 6) & 0x3f),
+        0x80 | (charcode & 0x3f)
+      );brave
+    }
+  }
+  return utf8;
+}
