+++ src/inspectors/ensureHashesEqual.js
@@ -1,0 +1,11 @@
+import VerifierError from '../models/verifierError';
+import * as SUB_STEPS from '../constants/verificationSubSteps';
+
+export default function ensureHashesEqual (actual, expected) {
+  if (actual !== expected) {
+    throw new VerifierError(
+      SUB_STEPS.compareHashes,
+      'Computed hash does not match remote hash'
+    );
+  }
+}
