+++ src/inspectors/index.js
@@ -1,0 +1,23 @@
+import computeLocalHash from './computeLocalHash';
+import ensureHashesEqual from './ensureHashesEqual';
+import ensureIssuerSignature from './ensureIssuerSignature';
+import ensureMerkleRootEqual from './ensureMerkleRootEqual';
+import ensureNotExpired from './ensureNotExpired';
+import ensureNotRevokedByList from './ensureNotRevokedByList';
+import ensureNotRevokedBySpentOutput from './ensureNotRevokedBySpentOutput';
+import ensureValidIssuingKey from './ensureValidIssuingKey';
+import ensureValidReceipt from './ensureValidReceipt';
+import isTransactionIdValid from './isTransactionIdValid';
+
+export {
+  computeLocalHash,
+  ensureHashesEqual,
+  ensureIssuerSignature,
+  ensureMerkleRootEqual,
+  ensureNotExpired,
+  ensureNotRevokedByList,
+  ensureNotRevokedBySpentOutput,
+  ensureValidIssuingKey,
+  ensureValidReceipt,
+  isTransactionIdValid
+};
